refactor: rename textEditor to inputEditor to make it more common name
