+++ packages/common/src/editors/editors.index.ts
@@ -3,7 +3,7 @@ import { CheckboxEditor } from './checkboxEditor';
 import { DateEditor } from './dateEditor';
 import { DualInputEditor } from './dualInputEditor';
 import { FloatEditor } from './floatEditor';
-import { InputEditor } from './textEditor';
+import { InputEditor } from './inputEditor';
 import { IntegerEditor } from './integerEditor';
 import { LongTextEditor } from './longTextEditor';
 import { MultipleSelectEditor } from './multipleSelectEditor';
@@ -23,10 +23,10 @@ export const Editors = {
   /** Dual Input Editor, default input type is text but it could be (integer/float/number/password/text) */
   dualInput: DualInputEditor,
 
-  /** Float Number Editor */
+  /** Float Number Editor using an input of type "number" */
   float: FloatEditor,
 
-  /** Integer Editor */
+  /** Integer Number Editor using an input of type "number" */
   integer: IntegerEditor,
 
   /** Long Text Editor (uses a textarea) */
@@ -38,9 +38,9 @@ export const Editors = {
   /** Single Select editor (which uses 3rd party lib "multiple-select.js") */
   singleSelect: SingleSelectEditor,
 
-  /** Slider Editor */
+  /** Slider Editor using an input of type "range" */
   slider: SliderEditor,
 
-  /** Text Editor */
+  /** Text Editor using an input of type "text" */
   text: InputEditor
 };
