+++ packages/common/src/editors/integerEditor.ts
@@ -1,7 +1,7 @@
 import { KeyCode } from '../enums/index';
 import { EditorArguments, EditorValidationResult } from '../interfaces/index';
 import { integerValidator } from '../editorValidators/integerValidator';
-import { InputEditor } from './textEditor';
+import { InputEditor } from './inputEditor';
 import { getDescendantProperty } from '../services/utilities';
 
 export class IntegerEditor extends InputEditor {
